backend/pyproject.toml
[project]
name = "mattilda-backend"
version = "0.1.0"
description = "FastAPI backend for Mattilda take-home exercise"
requires-python = ">=3.12"
dependencies = [
  "fastapi",
  "uvicorn[standard]",
  "sqlalchemy",
  "psycopg2-binary",
  "pydantic-settings",
  "redis",
  "pytest",
  "pytest-cov",
  "httpx",
  "python-jose[cryptography]",
  "passlib",
  "python-multipart",
  "email-validator",
  "alembic",
]
